+++ CHANGELOG.md
@@ -11,6 +11,10 @@ Note that since we don't clearly distinguish between a public and private interf
     - Fix camera stutter for "camera spin" animation
 - Add partial charge parsing support for MOL/SDF files (thanks @ptourlas)
 - [Breaking] Cleaner looking ``MembraneOrientationVisuals`` defaults
+- [Breaking] Add rock animation to trackball controls
+    - Add ``animate`` to ``TrackballControlsParams``, remove ``spin`` and ``spinSpeed``
+    - Add ``animate`` to ``SimpleSettingsParams``, remove ``spin``
+- Add "camera rock" state animation
 
 ## [v3.0.0-dev.9] - 2022-01-09
 
@@ -22,9 +26,6 @@ Note that since we don't clearly distinguish between a public and private interf
     - Add ``key`` field for mapping to source data
     - Fix assignment of bonds with unphysical length
 - Fix label/stats of single atom selection in multi-chain units
-- [Breaking] Add rock animation to trackball controls
-    - Add ``animate`` to ``TrackballControlsParams``, remove ``spin`` and ``spinSpeed``
-    - Add ``animate`` to ``SimpleSettingsParams``, remove ``spin``
 
 ## [v3.0.0-dev.8] - 2021-12-31